+++ mobile/android/android-components/components/browser/icons/src/main/java/mozilla/components/browser/icons/BrowserIcons.kt
@@ -231,6 +231,18 @@ class BrowserIcons(
         }
     }
 
+    /**
+     * Clears all icons and metadata from disk and memory.
+     *
+     * This will clear the default disk and memory cache that is used by the default configuration.
+     * If custom [IconLoader] and [IconProcessor] instances with a custom storage are provided to
+     * [BrowserIcons] then the calling app is responsible for clearing that data.
+     */
+    private fun clear() {
+        sharedDiskCache.clear(context)
+        sharedMemoryCache.clear()
+    }
+
     private suspend fun subscribeToUpdates(
         store: BrowserStore,
         flow: Flow<BrowserState>,