+++ mobile/android/android-components/docs/changelog.md
@@ -4,13 +4,21 @@ title: Changelog
 permalink: /changelog/
 ---
 
-# 52.0.0-SNAPSHOT (In Development)
+# 53.0.0-SNAPSHOT (In Development)
 
-* [Commits](https://github.com/mozilla-mobile/android-components/compare/v51.0.0...master)
-* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/111?closed=1)
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v52.0.0...master)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/112?closed=1)
 * [Dependencies](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Dependencies.kt)
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/master/buildSrc/src/main/java/Config.kt)
+
+# 52.0.0
+
+* [Commits](https://github.com/mozilla-mobile/android-components/compare/v51.0.0...v52.0.0)
+* [Milestone](https://github.com/mozilla-mobile/android-components/milestone/111?closed=1)
+* [Dependencies](https://github.com/mozilla-mobile/android-components/blob/v52.0.0/buildSrc/src/main/java/Dependencies.kt)
+* [Gecko](https://github.com/mozilla-mobile/android-components/blob/v52.0.0/buildSrc/src/main/java/Gecko.kt)
+* [Configuration](https://github.com/mozilla-mobile/android-components/blob/v52.0.0/buildSrc/src/main/java/Config.kt)
 
 * **browser-icons**
   * Added `BrowserIcons.clear()` to remove all saved data from disk and memory caches.